hyp_dlm/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hyp-dlm"
version = "0.1.0"
description = "HyP-DLM: Hypergraph Propagation with Dynamic Logic Modulation for Multi-hop QA"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
    "scikit-learn>=1.3",
    "sentence-transformers>=2.2",
    "spacy>=3.6",
    "hdbscan>=0.8.33",
    "faiss-cpu>=1.7",
    "networkx>=3.1",
    "pyyaml>=6.0",
    "openai>=1.0",
    "jellyfish>=1.0",
    "tqdm>=4.65",
    "rich>=13.0",
]

[project.optional-dependencies]
gliner = ["gliner>=0.2"]
gpu = ["faiss-gpu>=1.7"]
dev = ["pytest>=7.0", "pytest-cov>=4.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
